Fix download link location
